fixed scaling rounding erro

+++ jni/src/Application.cpp
@@ -133,10 +133,10 @@ int Application::startSDL()
 
   // Since the window size can be overriden, check what it is actually
   SDL_GetWindowSize(window, &global::viewport.w, &global::viewport.h);
-  global::stretch = V2f(global::viewport.w/WINDOW_DEFAULT_W,
-                        global::viewport.h/WINDOW_DEFAULT_H);
+  global::stretch = V2f(global::viewport.w / (float)WINDOW_DEFAULT_W,
+                        global::viewport.h / (float)WINDOW_DEFAULT_H);
   char buffer[16];
-  sprintf(buffer, "(%f, %f)", global::stretch.x, global::stretch.y);
+  sprintf(buffer, "(%.2f, %.2f)", global::stretch.x, global::stretch.y);
   LOG_I("Window scaling", buffer);
 
   // Create the OpenGL context for the window we just opened
